src/llm_bot/requirements.txt
python-telegram-bot
sqlalchemy
openai
json_repair
requests
asyncpg
loguru
fastapi
pydantic-settings
gunicorn
uvicorn
langchain
langchain-openai